requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-environ==0.11.2
django-ipware==6.0.3
django-rest-authemail==2.1.7
djangorestframework==3.14.0
idna==3.6
python-ipware==2.0.1
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
